Updated Get-WindowsAuditData script to utilise Test-RemoteConnection
Now this script will test the remote connectivity prior to attempting to run the audit. This gives a much more contextual error message about remote connectivity, and simplifies the code. In addition to this, removed the port option because it's fairly unused - most people don't even have WinRM configured let alone have it configured on SSL or a different port. Removed protocol too as we're now testing ourselves for this.