Added multiple layers of escaping to allow passwords which contain special chars
Unfortunately PSExec has some specialities around which characters are allowed in passwords, add to this CMD having its own escape sequences and PowerShell on top of that - some serious escaping had to be done in order to get around this problem. Thoroughly tested and working though, the only characters now that cannot be used in passwords is single/double quotes. If either of these is found in a password, a very clear exception is thrown