Finally resolved the application start issue inside the docker container
This was quite a pain, as it turns out dispite the fact we're using local port redirect to access the container locally, and ASP.NET is listening locally - this is not enough. Added * listener in the app start method to combat this issue. Also the .NET Directory.GetCurrentDirectory() method was returning an incorrect path due to the exe being started by fully qualified path from the dockerfile CMD statement, used a WORKDIR statement inside the dockerfile to set the working directory to the audit folder, then just called the exe by name in the CMD statement instead. Also changed the docker run options to better support this behaviour.