Added functionality to use distro-semantic scripting based on probe data
It's very important that we are able to run specific linux scripts against specific versions, as some functionality is not available in some distros. Added the ability to generate a distribution:version specific folder with a modifier to instruct the parser to get these scripts first when scanning a target matching the version+modifier. Also added a script to generate the specific folder based on a live host you're testing against.